[TASK] Adapt JavaScript entry point to module name convention
The first two path parts of scoped package identifiers
are similar to composer names: `@{vendor}/{package-name}`.
That means `ckeditor` is not actually a suitable package
name for this extension and is now replaced by the term
`ckeditor-wordbreak`. [1]

For the sake of this single purpose extension that means the
main entry point can be used as-is as a module specifier:
@sebkln/ckeditor-wordbreak

An alternative would have been to use the module identifier
`@sebkln/ckeditor-wordbreak/wordbreak.js`  but that's only
needed for extensions that ships multiple plugins.

[1] https://docs.typo3.org/m/typo3/reference-coreapi/main/en-us/ApiOverview/Backend/JavaScript/ES6/Index.html#configuration